+++ src/UiRenderer.cpp
@@ -33,8 +33,8 @@ void UiRenderer::operator()(Ptr<World> world) {
     world_ = world;
     rect_.x = 0;
     rect_.y = 0;
-    rect_.width = world->camera()->viewportWidth();
-    rect_.height = world->camera()->viewportHeight();
+    rect_.width = GLfloat(world->camera()->viewportWidth());
+    rect_.height = GLfloat(world->camera()->viewportHeight());
     operator()(world->ui());
 
     glDisable(GL_BLEND);
@@ -77,8 +77,8 @@ void UiRenderer::operator()(Ptr<Text> text) {
 
     // Pass the matrices to the vertex shader
     Ptr<Camera> camera = world_->camera();
-    GLfloat width = camera->viewportWidth();
-    GLfloat height = camera->viewportHeight();
+    GLfloat width = GLfloat(camera->viewportWidth());
+    GLfloat height = GLfloat(camera->viewportHeight());
     Matrix view = Matrix::ortho(0, width, height, 0, -1, 1);
     Matrix model = Matrix::translate(Vector(rect_.x, rect_.y, 0));
     Matrix scale = Matrix::scale(text->size(), -text->size(), 0);
